法律rag系统/requirements-docker.txt
# 核心 RAG 依赖
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
openai==1.82.0
pymilvus==2.5.9
pymongo==4.13.0
rank-bm25==0.2.2
requests==2.32.3
python-dotenv==1.1.0
jieba==0.42.1
numpy==1.26.4
pandas==2.2.3
tqdm==4.67.1
pydantic==2.10.6

# 文档处理
PyMuPDF==1.25.5
beautifulsoup4==4.13.4
lxml==5.4.0

# 评测
ragas==0.2.15
text2vec==1.3.5
datasets==3.6.0
sentence-transformers==4.1.0
torch==2.7.0
transformers==4.51.1

# 工具
huggingface-hub==0.32.3
fastapi==0.115.12
uvicorn==0.34.2
